Migrate FDML tests to munit
